Look for deps only where explicitly told to
Magically searching certain dirs is great for interactive use, but not
so much for a build system we want to be predictable (and where we can
afford to enumerate every dependency). Use a questionable hack to
defeat CMake's auto-search paths and then explicitly specify everything
else.

This should hopefully fix coexistance with MacPorts, and might fix some
other build failures we have been unable to track down but which might
be due to weird configurations people have their systems in.

Fixes #16