Properly specify ocaml location to CMake
The build scripts were updated to look for a new enough ocaml/ocamlc and
build out of third-party if one was not found; in doing so the necessary
defines here were changed, and not updated in the formula. This caused
the third-party ocaml to be unconditionally built, even though brew has
a new enough ocaml available, wasting a bunch of build time.

Fixes #76